% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_atomize.R
\name{mz_atomize}
\alias{mz_atomize}
\title{Parse a molecular formula into element and count}
\usage{
mz_atomize(mol)
}
\arguments{
\item{mol}{A string containing a molecular formula
(\emph{e.g.}, "C2H7NO3S"). Structural formulas containing parentheses are
not acceptable. Charges may be included, but the charge count should follow
the sign (\emph{e.g.}, "C10H16N5O13P3-3").}
}
\value{
A named vector of element counts.
}
\description{
\code{mz_atomize} separates a molecular formula into element and count,
returning a names list of element counts.
}
\examples{
mz_atomize("C5H8O5-")
mz_atomize("C2H7NO3S")
mz_atomize("C10H16N5O13P3-3")

}
