% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_iso_annotate.R
\name{mz_iso_annotate}
\alias{mz_iso_annotate}
\title{Annotate isotope table}
\usage{
mz_iso_annotate(molecule, polarity = "negative")
}
\arguments{
\item{molecule}{A string containing a molecular formula (\emph{e.g.}, "C2H7NO3S").
Structural formulas containing parentheses are not acceptable. Charges
may be included, but the charge count should follow the sign (\emph{e.g.},
"C10H16N5O13P3-3").}

\item{polarity}{Type of m/\emph{z} to return. Accepts one of "neutral", which
returns the monoisotopic mass; "positive", which returns the [M+H]+ mass;
or "negative", which returns the [M-H]- mass.}
}
\value{
A named list containing the molecule elements, stable isotopes, scan
polarity, and a tibble of annotated isotope information.
}
\description{
\code{mz_iso_annotate} generates a tibble enumerating all possible isotopes of
a molecule and annotates this with the mass or m/\emph{z} of each isotope
based on the polarity setting and the nominal mass shift of the isotope.
}
\examples{
mz_iso_annotate("C5H8O5")
mz_iso_annotate("C5H8O5", polarity = "negative")

}
