% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_iso_correct.R
\name{mz_iso_correct}
\alias{mz_iso_correct}
\title{Correct intensity vector for natural isotope abundance}
\usage{
mz_iso_correct(prob_matrix, v)
}
\arguments{
\item{prob_matrix}{Correction matrix output by \code{mz_iso_quant}}

\item{v}{Mass distribution vector of isotopologue intensities}
}
\value{
Non-negative mass distribution vector corrected for natural isotope abundance
}
\description{
Uses the \code{pracma::lsqlincon} function to correct the measured mass isotope
distribution for natural abundance with a lower bound of 0 for all corrected
values.
}
\examples{
intensity_vector <- c(5710407, 847435, 66688649, 3426642, 8222080, 99981, 0)
v <- intensity_vector/sum(intensity_vector)
prob_matrix <- mz_iso_quant("C6H14O12P2")$prob_matrix
mz_iso_correct(prob_matrix, v)

}
