% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_iso_list.R
\name{mz_iso_list}
\alias{mz_iso_list}
\title{Generate a matrix of stable isotopes}
\usage{
mz_iso_list(mol)
}
\arguments{
\item{mol}{A string containing a molecular formula (\emph{e.g.}, "C2H7NO3S").
Structural formulas containing parentheses are not acceptable. Charges
may be included, but the charge count should follow the sign (\emph{e.g.},
"C10H16N5O13P3-3").}
}
\value{
A matrix of isotope combinations
}
\description{
\code{mz_iso_list} generates a matrix enumerating all possible isotopes
of a molecule.
}
\examples{
mz_iso_list("C5H8O5")

}
