% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_iso_quant.R
\name{mz_iso_quant}
\alias{mz_iso_quant}
\title{Calculate isotope abundances}
\usage{
mz_iso_quant(
  mol,
  pol = "negative",
  tracer,
  purity = list(C = 1, H = 1, O = 1, N = 1, S = 1),
  ...
)
}
\arguments{
\item{isotope_matrix}{Matrix output of \code{mz_iso_tally}.}

\item{m}{Mass shift from the incorporation of a labed isotope.}

\item{lab}{Character vector of labed element.}
}
\value{
Vector of probabilities with length equal to the number of isotopes.
}
\description{
\code{mz_iso_quant} will calculate the probability matrix.
}
\examples{
z <- mz_iso_list("C3")

## Calculate the natural abundance.
mz_iso_calc_probs(z)

## Calculate the natural abundance assuming one labled carbon.
mz_iso_calc_probs(z, lab = "C", m = 1)

}
