% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzrtools.dev-package.R
\docType{package}
\name{mzrtools.dev-package}
\alias{mzrtools.dev}
\alias{mzrtools.dev-package}
\title{mzrtools.dev: Make Molecular Formulas Useful for Mass Spectrometry Applications}
\description{
Several tools that convert molecular formulas into useful values 
    for mass spectrometry applications. This includes parsing formulas, 
    calculating monoisotopic masses, and building chemical libraries for 
    targeted mass spectrometry applications. In addition, functions will 
    enumerate chemical isotopes, calculate the isotopic probability 
    distributions, and perform correction for natural isotope abundance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wmoldham/mzrtools.dev}
  \item Report bugs at \url{https://github.com/wmoldham/mzrtools.dev/issues}
}

}
\author{
\strong{Maintainer}: William Oldham \email{wm.oldham@gmail.com} (\href{https://orcid.org/0000-0003-3029-4866}{ORCID})

}
\keyword{internal}
