% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_iso_resolve.R
\name{mz_iso_resolve}
\alias{mz_iso_resolve}
\title{Filter unlabeled resolved isotopes}
\usage{
mz_iso_resolve(mol, tracer, pol = "negative", ...)
}
\arguments{
\item{mol}{A string containing a molecular formula (\emph{e.g.}, "C2H7NO3S").
Structural formulas containing parentheses are not acceptable. Charges
may be included, but the charge count should follow the sign (\emph{e.g.},
"C10H16N5O13P3-3").}

\item{tracer}{A vector of elements labeled by the tracer.}

\item{pol}{Type of m/\emph{z} to return. Accepts one of "neutral", which
returns the monoisotopic mass; "positive", which returns the [M+H]+ mass;
or "negative", which returns the [M-H]- mass.}

\item{...}{Passes the arguments nominal_resolution (default 70,000) and
nominal_mz (default 200) used to determine the delta m/\emph{z} below
which peaks are considered unresolved.}
}
\value{
A tibble containing the unit mass shift (\code{m}), group, mass or
m/\emph{z}, and isotope composition.
}
\description{
\code{mz_iso_resolve} will return the molecular isotopes of interest
(\emph(e.g.), M0 or label-containing) and those isotopes that cannot be
resolved from these targets by the mass spectrometer.
}
\examples{
mz_iso_resolve("C5H8O5")
mz_iso_resolve("C5H8O5", tracer = c("C", "H"))

}
